requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.12.2
psycopg2==2.9.3
sqlparse==0.4.2
